+++ iOS/pet2share/Controllers/Profile/PetProfileVC.m
@@ -157,7 +157,7 @@ static NSString * const kLeftIconImageName      = @"icon-arrowback";
 - (void)onReceiveSuccess:(NSArray *)objects
 {
     _isRequesting = NO;
-    fTRACE(@"Number of Objects: %ld", objects.count);
+    fTRACE(@"Number of Objects: %ld", (long)objects.count);
     if (self.refreshControl.isRefreshing || _pageNumber == 1)
     {
         [self.items removeAllObjects];
